requirements.txt
langchain
langchain_groq
langchain_core

python-dotenv
pinecone
openai

flask
flask-cors